refactor(agent): Move sqlite client to utils
