+++ .gitignore
@@ -179,7 +179,7 @@ infra/*/*.tfstate
 infra/*/*.tfstate.backup
 infra/*/.terraform*
 infra/*/*.plan
-policies.db
+*.db
 agent/resources
 agent/app/environment/resources
 agent/app/environment/seller/resources
