+++ agent/app/policies/negotiation_thread.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from typing import Any, Dict, List, Optional
 from dataclasses import dataclass
 
-from app.policies.sqlite_client import SQLiteClient
+from app.utils.sqlite_client import SQLiteClient
 from app.utils.config import CONFIG
 
 logger = logging.getLogger(__name__)
